ㅇㅁㄴㅀㅈㄷㅎㅈㅁㅎ

+++ app/src/main/java/com/example/calcal/modelDTO/RouteAndTimeDTO.kt
@@ -3,5 +3,5 @@ package com.example.calcal.modelDTO
 data class RouteAndTimeDTO(
     var latitude: Double,
     var longitude: Double,
-    var recordTime: Long
+    var recordTime: Double
 )
